--- a/src/Plots.jl
+++ b/src/Plots.jl
@@ -245,4 +245,6 @@ end
 
 const CURRENT_BACKEND = CurrentBackend(:none)
 
+include("precompile_includer.jl")
+
 end # module