I started looking for invalidations trying to follow @timholy s recommendations - don't remember from which issue comment. So, I deactivated precompilation for Plots, did something along the lines of
```julia
timing = @snoopi tmin = 0.01 include(joinpath(@__DIR__, "precompile_script.jl"))

mtimes = Dict{Method,Float64}()
for (t, mi) in timing
    meth = mi.def
    if isa(meth, Method)
        mtimes[meth] = t + get(mtimes, meth, 0.0)
    end
end

tinfm = [(v, k) for (k, v) in mtimes];
sort!(tinfm; by=first)
```
and looked at the timings. First I realized that `warn_on_recipe_aliases!`, which is really just a small utility to inform users, took a lot of time:
```
(1.463343858718872, warn_on_recipe_aliases!(plt::Plots.Plot, v::AbstractArray{T,1} where T, recipe_type, args...) in Plots at /home/daniel/.julia/dev/Plots/src/pipeline.jl:21)
```
So I moved this to RecipesPipeline and changed the signatures a little bit.
From there the only thing I did was looking at other timings, adding some `@nospecialize` (where I believed appropriate in RecipesBase, RecipesPipeline and Plots) and running the above script again for several times.
I am not really that knowledgable about this so it would be great if @timholy - if he can find a little time or someone with similar expertise - could have a quick look at the `@nospecialize` statements, I added, and see if they do make sense.

Anyway, together with https://github.com/JuliaPlots/RecipesBase.jl/pull/75 and https://github.com/JuliaPlots/RecipesPipeline.jl/pull/67 this achieves the following improvements for the time to first plot (`@time (using Plots; display(plot(rand(10))))`) compared to the current Plots release on my PC (3 runs each):

#### Julia 1.5.2
Current Plots release:
```
19.840764 seconds (29.85 M allocations: 1.595 GiB, 3.63% gc time)
18.705377 seconds (29.85 M allocations: 1.596 GiB, 3.09% gc time)
18.364920 seconds (29.85 M allocations: 1.596 GiB, 3.14% gc time)
```
Now:
```
15.110489 seconds (22.87 M allocations: 1.257 GiB, 2.98% gc time)
15.573129 seconds (22.88 M allocations: 1.257 GiB, 2.80% gc time)
15.356651 seconds (22.88 M allocations: 1.257 GiB, 2.78% gc time)
```
#### Julia master
Current Plots release:
```
11.527787 seconds (18.64 M allocations: 1.148 GiB, 3.80% gc time, 74.31% compilation time)
12.837436 seconds (18.64 M allocations: 1.148 GiB, 3.73% gc time, 72.42% compilation time)
12.814848 seconds (18.64 M allocations: 1.148 GiB, 3.73% gc time, 75.42% compilation time)
```
Now:
```
10.526915 seconds (17.02 M allocations: 1.047 GiB, 4.06% gc time, 72.13% compilation time)
12.133592 seconds (17.02 M allocations: 1.047 GiB, 3.72% gc time, 75.20% compilation time)
12.504285 seconds (17.02 M allocations: 1.047 GiB, 3.50% gc time, 73.42% compilation time)
```
Especially on Julia master the improvements in time are marginal, but I suppose it's a small step in the right direction.