@timholy 2020-11-07
Sorry for the delay in looking at this. I didn't see any problems with this, looks like a good change to me.